chore(shared-influx-pieces): move functions in lib folder
